+++ R/RcppExports.R
@@ -126,6 +126,10 @@ inner_join_impl <- function(x, y, by_x, by_y, aux_x, aux_y, na_match) {
     .Call(`_dplyr_inner_join_impl`, x, y, by_x, by_y, aux_x, aux_y, na_match)
 }
 
+nest_join_impl <- function(x, y, by_x, by_y, aux_x, aux_y, na_match, yname) {
+    .Call(`_dplyr_nest_join_impl`, x, y, by_x, by_y, aux_x, aux_y, na_match, yname)
+}
+
 left_join_impl <- function(x, y, by_x, by_y, aux_x, aux_y, na_match) {
     .Call(`_dplyr_left_join_impl`, x, y, by_x, by_y, aux_x, aux_y, na_match)
 }